+++ mitmproxy/proxy/layers/http/__init__.py
@@ -321,6 +321,8 @@ class HttpStream(layer.Layer):
             else:
                 chunks = [event.data]
             for chunk in chunks:
+                if self.context.options.store_streamed_bodies:
+                    self.request_body_buf += chunk
                 yield SendHttp(RequestData(self.stream_id, chunk), self.context.server)
         elif isinstance(event, RequestTrailers):
             # we don't do anything further here, we wait for RequestEndOfMessage first to trigger the request hook.
@@ -333,10 +335,15 @@ class HttpStream(layer.Layer):
                 elif isinstance(chunks, bytes):
                     chunks = [chunks]
                 for chunk in chunks:
+                    if self.context.options.store_streamed_bodies:
+                        self.request_body_buf += chunk
                     yield SendHttp(
                         RequestData(self.stream_id, chunk), self.context.server
                     )
 
+            if self.context.options.store_streamed_bodies:
+                self.flow.request.data.content = bytes(self.request_body_buf)
+                self.request_body_buf.clear()
             self.flow.request.timestamp_end = time.time()
             yield HttpRequestHook(self.flow)
             self.client_state = self.state_done
@@ -444,6 +451,8 @@ class HttpStream(layer.Layer):
             else:
                 chunks = [event.data]
             for chunk in chunks:
+                if self.context.options.store_streamed_bodies:
+                    self.response_body_buf += chunk
                 yield SendHttp(ResponseData(self.stream_id, chunk), self.context.client)
         elif isinstance(event, ResponseTrailers):
             self.flow.response.trailers = event.trailers
@@ -456,9 +465,14 @@ class HttpStream(layer.Layer):
                 elif isinstance(chunks, bytes):
                     chunks = [chunks]
                 for chunk in chunks:
+                    if self.context.options.store_streamed_bodies:
+                        self.response_body_buf += chunk
                     yield SendHttp(
                         ResponseData(self.stream_id, chunk), self.context.client
                     )
+            if self.context.options.store_streamed_bodies:
+                self.flow.response.data.content = bytes(self.response_body_buf)
+                self.response_body_buf.clear()
             yield from self.send_response(already_streamed=True)
 
     @expect(ResponseData, ResponseTrailers, ResponseEndOfMessage)