+++ mitmproxy/proxy/server.py
@@ -507,6 +507,12 @@ if __name__ == "__main__":  # pragma: no cover
     opts = moptions.Options()
     # options duplicated here to simplify testing setup
     opts.add_option(
+        "store_streamed_bodies",
+        bool,
+        False,
+        "Store HTTP request and response bodies when streamed.",
+    )
+    opts.add_option(
         "connection_strategy",
         str,
         "lazy",